move isRecurringActive to config
